Add "free university lecture signup manager" code from 2011
This was a spinoff over the WhispyForum engine to handle a school
project's handling where individual students could sign up for
individual lessons (given the participant limit hasn't been reached
yet).

The history of this development was lost to the sands of time with the
SVN-2-Git migration and the original host of the project, Google Code,
having shut down since.